chore: refactor middleware handler generation
