+++ smithy-swift-codegen/src/main/kotlin/software/amazon/smithy/swift/codegen/integration/middlewares/handlers/HttpHeaderMiddleware.kt
@@ -3,18 +3,25 @@
  * SPDX-License-Identifier: Apache-2.0.
  */
 
-package software.amazon.smithy.swift.codegen.integration
+package software.amazon.smithy.swift.codegen.integration.middlewares.handlers
 
 import software.amazon.smithy.codegen.core.Symbol
 import software.amazon.smithy.model.knowledge.HttpBinding
 import software.amazon.smithy.model.knowledge.HttpBindingIndex
 import software.amazon.smithy.model.shapes.CollectionShape
 import software.amazon.smithy.model.shapes.MemberShape
+import software.amazon.smithy.model.shapes.OperationShape
 import software.amazon.smithy.model.traits.TimestampFormatTrait
 import software.amazon.smithy.swift.codegen.ClientRuntimeTypes
 import software.amazon.smithy.swift.codegen.Middleware
+import software.amazon.smithy.swift.codegen.MiddlewareGenerator
+import software.amazon.smithy.swift.codegen.SwiftDependency
 import software.amazon.smithy.swift.codegen.SwiftTypes
 import software.amazon.smithy.swift.codegen.SwiftWriter
+import software.amazon.smithy.swift.codegen.integration.HttpBindingDescriptor
+import software.amazon.smithy.swift.codegen.integration.HttpBindingResolver
+import software.amazon.smithy.swift.codegen.integration.ProtocolGenerator
+import software.amazon.smithy.swift.codegen.integration.formatHeaderOrQueryValue
 import software.amazon.smithy.swift.codegen.integration.steps.OperationSerializeStep
 import software.amazon.smithy.swift.codegen.model.defaultValue
 import software.amazon.smithy.swift.codegen.model.isBoxed
@@ -33,7 +40,36 @@ class HttpHeaderMiddleware(
 
     private val bindingIndex = HttpBindingIndex.of(ctx.model)
     override val typeName = "${inputSymbol.name}HeadersMiddleware"
+    companion object {
+        fun renderHeaderMiddleware(
+            ctx: ProtocolGenerator.GenerationContext,
+            op: OperationShape,
+            httpBindingResolver: HttpBindingResolver,
+            defaultTimestampFormat: TimestampFormatTrait.Format
+        ) {
+            val requestBindings = httpBindingResolver.requestBindings(op)
+            val headerBindings = requestBindings
+                .filter { it.location == HttpBinding.Location.HEADER }
+                .sortedBy { it.memberName }
+            val prefixHeaderBindings = requestBindings
+                .filter { it.location == HttpBinding.Location.PREFIX_HEADERS }
 
+            val inputSymbol = MiddlewareHandlerUtils.inputSymbol(ctx, op)
+            val outputSymbol = MiddlewareHandlerUtils.outputSymbol(ctx, op)
+            val outputErrorSymbol = MiddlewareHandlerUtils.outputErrorSymbol(op)
+            val rootNamespace = MiddlewareHandlerUtils.rootNamespace(ctx)
+
+            val headerMiddlewareSymbol = Symbol.builder()
+                .definitionFile("./$rootNamespace/models/${inputSymbol.name}+HeaderMiddleware.swift")
+                .name(inputSymbol.name)
+                .build()
+            ctx.delegator.useShapeWriter(headerMiddlewareSymbol) { writer ->
+                writer.addImport(SwiftDependency.CLIENT_RUNTIME.target)
+                val headerMiddleware = HttpHeaderMiddleware(writer, ctx, inputSymbol, outputSymbol, outputErrorSymbol, headerBindings, prefixHeaderBindings, defaultTimestampFormat)
+                MiddlewareGenerator(writer, headerMiddleware).generate()
+            }
+        }
+    }
     override fun generateMiddlewareClosure() {
         generateHeaders()
         generatePrefixHeaders()
